Tear kv.h into multiple files
